requirements.txt
# Minimal runtime requirements for running the Flask app
flask
flask-cors
python-dotenv
requests

# Dev tools
autopep8

# Optional / project-specific dependencies (commented out to avoid heavy builds during local setup)
# pandas
# numpy
# requests
# openai
# Pillow
# aiortc
